refactor : 중복된 로직 제거
